Refactor: Moved all software boxes into folders
